Link RenderTexture2
