+++ RenderTexture1/scenes/sce_stage.ts
@@ -8,7 +8,6 @@ import {World} from "../world.js";
 
 export function scene_stage(game: Game) {
     game.World = new World();
-    game.Cameras = [];
     game.ViewportResized = true;
 
     // Main Camera.
