+++ RenderTexture2/blueprints/blu_camera_main.ts
@@ -1,4 +1,4 @@
-import {camera_display} from "../components/com_camera_display.js";
+import {camera_display_perspective} from "../components/com_camera.js";
 import {Blueprint} from "../core.js";
 import {Game} from "../game.js";
 
@@ -8,7 +8,7 @@ export function blueprint_camera_main(game: Game): Blueprint {
         Children: [
             {
                 Rotation: [0, 1, 0, 0],
-                Using: [camera_display(1, 0.1, 1000, [0.9, 0.9, 0.9, 1])],
+                Using: [camera_display_perspective(1, 0.1, 1000, [0.9, 0.9, 0.9, 1])],
             },
         ],
     };
