--- RenderTexture2/components/com_render.ts
@@ -1,7 +1,0 @@
-import {RenderTextured} from "./com_render_textured.js";
-
-export type Render = RenderTextured;
-
-export const enum RenderKind {
-    Textured,
-}
