+++ RenderTexture2/components/com_render2.ts
@@ -1,0 +1,1 @@
+../../core/components/com_render2.ts
