--- RenderTexture2/components/com_render_textured.ts
@@ -1,68 +1,0 @@
-import {Material, Mesh} from "../../common/material.js";
-import {GL_ARRAY_BUFFER, GL_CW, GL_ELEMENT_ARRAY_BUFFER, GL_FLOAT} from "../../common/webgl.js";
-import {Entity, Game} from "../game.js";
-import {TexturedLayout} from "../materials/layout_textured.js";
-import {Has} from "../world.js";
-import {RenderKind} from "./com_render.js";
-
-export interface RenderTextured {
-    Kind: RenderKind.Textured;
-    Material: Material<TexturedLayout>;
-    Mesh: Mesh;
-    FrontFace: GLenum;
-    Vao: WebGLVertexArrayObject;
-    Texture: WebGLTexture;
-}
-
-let vaos: WeakMap<Mesh, WebGLVertexArrayObject> = new WeakMap();
-
-export function render_textured(
-    material: Material<TexturedLayout>,
-    mesh: Mesh,
-    texture: WebGLTexture
-) {
-    return (game: Game, entity: Entity) => {
-        if (!vaos.has(mesh)) {
-            // We only need to create the VAO once.
-            let vao = game.Gl.createVertexArray()!;
-            game.Gl.bindVertexArray(vao);
-
-            game.Gl.bindBuffer(GL_ARRAY_BUFFER, mesh.VertexBuffer);
-            game.Gl.enableVertexAttribArray(material.Locations.VertexPosition);
-            game.Gl.vertexAttribPointer(
-                material.Locations.VertexPosition,
-                3,
-                GL_FLOAT,
-                false,
-                0,
-                0
-            );
-
-            game.Gl.bindBuffer(GL_ARRAY_BUFFER, mesh.TexCoordBuffer);
-            game.Gl.enableVertexAttribArray(material.Locations.VertexTexCoord);
-            game.Gl.vertexAttribPointer(
-                material.Locations.VertexTexCoord,
-                2,
-                GL_FLOAT,
-                false,
-                0,
-                0
-            );
-
-            game.Gl.bindBuffer(GL_ELEMENT_ARRAY_BUFFER, mesh.IndexBuffer);
-
-            game.Gl.bindVertexArray(null);
-            vaos.set(mesh, vao);
-        }
-
-        game.World.Signature[entity] |= Has.Render;
-        game.World.Render[entity] = {
-            Kind: RenderKind.Textured,
-            Material: material,
-            Mesh: mesh,
-            FrontFace: GL_CW,
-            Vao: vaos.get(mesh)!,
-            Texture: texture,
-        };
-    };
-}
