--- RenderTexture2/materials/layout_textured.ts
@@ -1,10 +1,0 @@
-export interface TexturedLayout {
-    // Uniforms
-    Pv: WebGLUniformLocation;
-    World: WebGLUniformLocation;
-    Self: WebGLUniformLocation;
-    Sampler: WebGLUniformLocation;
-    // Attributes
-    VertexPosition: GLint;
-    VertexTexCoord: GLint;
-}
