+++ RenderTexture2/scenes/sce_stage.ts
@@ -1,6 +1,6 @@
 import {blueprint_camera_main} from "../blueprints/blu_camera_main.js";
 import {blueprint_camera_minimap} from "../blueprints/blu_camera_minimap.js";
-import {render_textured} from "../components/com_render_textured.js";
+import {render_textured} from "../components/com_render2.js";
 import {rotate} from "../components/com_rotate.js";
 import {instantiate} from "../core.js";
 import {Game} from "../game.js";
@@ -8,7 +8,6 @@ import {World} from "../world.js";
 
 export function scene_stage(game: Game) {
     game.World = new World();
-    game.Cameras = [];
     game.ViewportResized = true;
 
     // Main Camera.
