+++ RenderTexture2/systems/sys_render2.ts
@@ -1,0 +1,1 @@
+../../core/systems/sys_render2.ts
