+++ core/components/com_render2.ts
@@ -3,14 +3,16 @@ import {Vec4} from "../../common/math.js";
 import {GL_ARRAY_BUFFER, GL_CW, GL_ELEMENT_ARRAY_BUFFER, GL_FLOAT} from "../../common/webgl.js";
 import {DiffuseLayout} from "../../materials/layout_diffuse.js";
 import {SpecularLayout} from "../../materials/layout_specular.js";
+import {TexturedLayout} from "../../materials/layout_textured.js";
 import {Entity, Game} from "../game.js";
 import {Has, World} from "../world.js";
 
-export type Render = RenderDiffuse | RenderSpecular;
+export type Render = RenderDiffuse | RenderSpecular | RenderTextured;
 
 export const enum RenderKind {
     Diffuse,
     Specular,
+    Textured,
 }
 
 interface Game2 extends Game {
@@ -128,6 +130,68 @@ export function render_specular(
             ColorDiffuse: color_diffuse,
             ColorSpecular: color_specular,
             Shininess: shininess,
+        };
+    };
+}
+
+export interface RenderTextured {
+    readonly Kind: RenderKind.Textured;
+    readonly Material: Material<TexturedLayout>;
+    readonly Mesh: Mesh;
+    readonly FrontFace: GLenum;
+    readonly Vao: WebGLVertexArrayObject;
+    Texture: WebGLTexture;
+}
+
+let render_textured_vaos: WeakMap<Mesh, WebGLVertexArrayObject> = new WeakMap();
+
+export function render_textured(
+    material: Material<TexturedLayout>,
+    mesh: Mesh,
+    texture: WebGLTexture
+) {
+    return (game: Game2, entity: Entity) => {
+        if (!render_textured_vaos.has(mesh)) {
+            // We only need to create the VAO once.
+            let vao = game.Gl.createVertexArray()!;
+            game.Gl.bindVertexArray(vao);
+
+            game.Gl.bindBuffer(GL_ARRAY_BUFFER, mesh.VertexBuffer);
+            game.Gl.enableVertexAttribArray(material.Locations.VertexPosition);
+            game.Gl.vertexAttribPointer(
+                material.Locations.VertexPosition,
+                3,
+                GL_FLOAT,
+                false,
+                0,
+                0
+            );
+
+            game.Gl.bindBuffer(GL_ARRAY_BUFFER, mesh.TexCoordBuffer);
+            game.Gl.enableVertexAttribArray(material.Locations.VertexTexCoord);
+            game.Gl.vertexAttribPointer(
+                material.Locations.VertexTexCoord,
+                2,
+                GL_FLOAT,
+                false,
+                0,
+                0
+            );
+
+            game.Gl.bindBuffer(GL_ELEMENT_ARRAY_BUFFER, mesh.IndexBuffer);
+
+            game.Gl.bindVertexArray(null);
+            render_textured_vaos.set(mesh, vao);
+        }
+
+        game.World.Signature[entity] |= Has.Render;
+        game.World.Render[entity] = {
+            Kind: RenderKind.Textured,
+            Material: material,
+            Mesh: mesh,
+            FrontFace: GL_CW,
+            Vao: render_textured_vaos.get(mesh)!,
+            Texture: texture,
         };
     };
 }
