+++ core/systems/sys_render2.ts
@@ -3,12 +3,21 @@ import {
     GL_COLOR_BUFFER_BIT,
     GL_DEPTH_BUFFER_BIT,
     GL_FRAMEBUFFER,
+    GL_TEXTURE0,
+    GL_TEXTURE_2D,
     GL_UNSIGNED_SHORT,
 } from "../../common/webgl.js";
 import {DiffuseLayout} from "../../materials/layout_diffuse.js";
 import {SpecularLayout} from "../../materials/layout_specular.js";
+import {TexturedLayout} from "../../materials/layout_textured.js";
 import {CameraDisplay, CameraEye, CameraFramebuffer, CameraKind} from "../components/com_camera.js";
-import {Render, RenderDiffuse, RenderKind, RenderSpecular} from "../components/com_render2.js";
+import {
+    Render,
+    RenderDiffuse,
+    RenderKind,
+    RenderSpecular,
+    RenderTextured,
+} from "../components/com_render2.js";
 import {Transform} from "../components/com_transform.js";
 import {Game} from "../game.js";
 import {Has, World} from "../world.js";
@@ -70,6 +79,9 @@ function render(game: Game2, eye: CameraEye, current_target?: WebGLTexture) {
                     case RenderKind.Specular:
                         use_specular(game, render.Material, eye);
                         break;
+                    case RenderKind.Textured:
+                        use_textured(game, render.Material, eye);
+                        break;
                 }
             }
 
@@ -84,6 +96,13 @@ function render(game: Game2, eye: CameraEye, current_target?: WebGLTexture) {
                     break;
                 case RenderKind.Specular:
                     draw_specular(game, transform, render);
+                    break;
+                case RenderKind.Textured:
+                    // Prevent feedback loop between the active render target
+                    // and the texture being rendered.
+                    if (render.Texture !== current_target) {
+                        draw_textured(game, transform, render);
+                    }
                     break;
             }
         }
@@ -120,6 +139,23 @@ function draw_specular(game: Game2, transform: Transform, render: RenderSpecular
     game.Gl.uniform4fv(render.Material.Locations.ColorDiffuse, render.ColorDiffuse);
     game.Gl.uniform4fv(render.Material.Locations.ColorSpecular, render.ColorSpecular);
     game.Gl.uniform1f(render.Material.Locations.Shininess, render.Shininess);
+    game.Gl.bindVertexArray(render.Vao);
+    game.Gl.drawElements(render.Material.Mode, render.Mesh.IndexCount, GL_UNSIGNED_SHORT, 0);
+    game.Gl.bindVertexArray(null);
+}
+
+function use_textured(game: Game2, material: Material<TexturedLayout>, eye: CameraEye) {
+    game.Gl.useProgram(material.Program);
+    game.Gl.uniformMatrix4fv(material.Locations.Pv, false, eye.Pv);
+}
+
+function draw_textured(game: Game2, transform: Transform, render: RenderTextured) {
+    game.Gl.uniformMatrix4fv(render.Material.Locations.World, false, transform.World);
+
+    game.Gl.activeTexture(GL_TEXTURE0);
+    game.Gl.bindTexture(GL_TEXTURE_2D, render.Texture);
+    game.Gl.uniform1i(render.Material.Locations.Sampler, 0);
+
     game.Gl.bindVertexArray(render.Vao);
     game.Gl.drawElements(render.Material.Mode, render.Mesh.IndexCount, GL_UNSIGNED_SHORT, 0);
     game.Gl.bindVertexArray(null);
